backend/requirements.txt
Django>=3.0.8,<3.1.0
djangorestframework>=3.11.0,<3.12.0
djangorestframework-simplejwt>=4.4.0,<4.5.0
psycopg2>=2.8.0,<2.9.0
pytest-asyncio>=0.14.0,<0.20.0
pytest-django>=3.9.0,<3.10.0
flake8==3.8.3
black==19.10b0
